Create MVC microservice for caching system.
Add two GET endpoints to retrieve journey(s) related to a user.
The cached data is stored as String-String pair, where the key is a user ID string, and the value is a list of Journey objects that has been stringified.
Add related tests for both endpoints.
Add README with instructions on how to run.
Add Notes file to clarify some thought processes.